FIX: report generating and billing process

+++ src/main/java/com/elstele/bill/form/CallForm.java
@@ -10,7 +10,7 @@ public class CallForm {
     private String numberA;
     private String numberB;
     private Date startTime;
-    private Long duration;
+    private Integer duration;
     private String aonKat;
     private String dvoCodeA;
     private String dvoCodeB;
@@ -89,11 +89,11 @@ public class CallForm {
         this.startTime = startTime;
     }
 
-    public Long getDuration() {
+    public Integer getDuration() {
         return duration;
     }
 
-    public void setDuration(Long duration) {
+    public void setDuration(Integer duration) {
         this.duration = duration;
     }
